Move catalog types to separate file
