+++ src/lsdb/core/search/index_search.py
@@ -8,7 +8,7 @@ from hipscat.catalog.index.index_catalog import IndexCatalog
 from lsdb.core.search.abstract_search import AbstractSearch
 
 if TYPE_CHECKING:
-    from lsdb.loaders.hipscat.abstract_catalog_loader import HCCatalogTypeVar
+    from lsdb.types import HCCatalogTypeVar
 
 
 class IndexSearch(AbstractSearch):
