+++ src/lsdb/types.py
@@ -1,7 +1,10 @@
-from typing import Dict, List, Tuple
+from typing import Dict, List, Tuple, TypeVar
 
+from hipscat.catalog.healpix_dataset.healpix_dataset import HealpixDataset as HCHealpixDataset
 from hipscat.pixel_math import HealpixPixel
 from typing_extensions import TypeAlias
+
+from lsdb.catalog.dataset.dataset import Dataset
 
 # Compute pixel map returns a tuple. The first element is
 # the number of data points within the HEALPix pixel, the
@@ -9,3 +12,6 @@ from typing_extensions import TypeAlias
 HealpixInfo: TypeAlias = Tuple[int, List[int]]
 
 DaskDFPixelMap = Dict[HealpixPixel, int]
+
+CatalogTypeVar = TypeVar("CatalogTypeVar", bound=Dataset)
+HCCatalogTypeVar = TypeVar("HCCatalogTypeVar", bound=HCHealpixDataset)
